requirements.txt
gevent==20.6.1
greenlet==0.4.16
crcmod==1.7
gevent-pika==0.2
future==0.18.2
pandas==0.25.3
sklearn==0.0
scikit-learn==0.20.3
pymodbus==2.3.0
python-nmap==0.6.1
pyYAML==5.2
fuzzywuzzy==0.17.0
pyshark==0.4.2.9
tftpy==0.8.0
urllib3==1.25.7
requests==2.22.0
pyTenable==1.0.0
python-Levenshtein==0.12.0
crccheck==0.6
flask
flask-restful
flask-cors
psycopg2-binary
cryptography
